Version 2.0
Added the ability to keep or close a tab when opening it in a proxy window.
Tweaked some of the way that preferences worked.
Cleaned up the interface to make it clearer how things worked.
Added a link to be able to get updates for the app.
Fixed a few minor bugs.
Using a new data storage space under a clearer container name.